Rename dynamics file, simplify so_rpy model, simplify models wrapper
