+++ crazyflow/dynamics/so_rpy_rotor_drag/dynamics.py
